Origin: equipitem packet added & added too the listener for game sync, facetiles added too prepareattack in combat, button 454 & 632 added to click packet listener

+++ src/main/java/com/aelous/model/entity/combat/CombatSpecial.java
@@ -350,7 +350,7 @@ public enum CombatSpecial {
             if (spec == CombatSpecial.GRANITE_MAUL) {
 
 
-                int specPercentage = (int) (player.getSpecialAttackPercentage());
+                int specPercentage = player.getSpecialAttackPercentage();
 
                 //Make sure the player has enough special attack
                 if (specPercentage < player.getCombatSpecial().getDrainAmount()) {
